# backend/openapi.yaml
openapi: 3.0.3
info:
  title: RavelGo API
  version: "0.1.0"
  description: >
    Unified REST API for the RavelGo rider, driver, and admin apps.
    Authenticated with Amazon Cognito access tokens; role checks are enforced
    via Cognito group membership (`Rider`, `Driver`, `Admin`) — see
    `src/middleware/auth.ts`. There is also a WebSocket endpoint
    (`wss://<host>/ws?token=<access token>`) for live trip-status and
    driver-location push, not representable in OpenAPI 3.0 — see
    `docs/realtime-architecture.md` for its wire protocol. Every REST
    endpoint below has an HTTP-polling equivalent, so a client never
    strictly needs the WebSocket connection.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /payouts:
    get:
      tags: [Payments]
      summary: All driver payouts (Admin)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated payouts
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /payouts/{id}:
    get:
      tags: [Payments]
      summary: One payout (Driver; own payouts only)
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Payout
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /payouts/history:
    get:
      tags: [Payments]
      summary: My own payout history (Driver)
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated payouts
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /payouts/banks:
    get:
      tags: [Payments]
      summary: Banks payouts can be sent to (Driver)
      responses:
        "200":
          description: Supported banks
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /payouts/bank-account:
    get:
      tags: [Payments]
      summary: My saved payout bank account (Driver)
      responses:
        "200":
          description: Bank account
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Payments]
      summary: Save/replace my payout bank account (Driver)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountNumber, bankCode]
              properties:
                accountNumber: { type: string }
                bankCode: { type: string }
      responses:
        "201":
          description: Bank account
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /payouts/calculate:
    post:
      tags: [Payments]
      summary: Preview what a driver is owed for a period (Admin, payouts:write)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driverId, periodStart, periodEnd]
              properties:
                driverId: { type: string }
                periodStart: { type: string }
                periodEnd: { type: string }
      responses:
        "200":
          description: Calculated payout
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /payouts/create:
    post:
      tags: [Payments]
      summary: Create a payout for a driver and period (Admin, payouts:write)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driverId, periodStart, periodEnd]
              properties:
                driverId: { type: string }
                periodStart: { type: string }
                periodEnd: { type: string }
      responses:
        "201":
          description: Payout
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /payouts/{id}/process:
    post:
      tags: [Payments]
      summary: Send a created payout to the provider (Admin, payouts:write)
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Payout
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /payouts/{id}/complete:
    post:
      tags: [Payments]
      summary: Mark a payout settled (Admin, payouts:write)
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Payout
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /payouts/{id}/fail:
    post:
      tags: [Payments]
      summary: Mark a payout failed (Admin, payouts:write)
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200":
          description: Payout
          content:
            application/json:
              schema: { type: object }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Cognito access token
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Validation or bad request error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Caller lacks the required Cognito group
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_ERROR
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_SERVER_ERROR
                - RATE_LIMIT_EXCEEDED
                - BAD_REQUEST
                - PAYMENT_ERROR
                - DATABASE_ERROR
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              nullable: true
              description: Field-level validation errors
              items:
                type: object
                properties:
                  field: { type: string, description: 'Field name or path (e.g., "email", "address.zipCode")' }
                  message: { type: string, description: Error message for this field }
                  code: { type: string, nullable: true, description: 'Error code (e.g., "too_small", "invalid_email")' }
            timestamp:
              type: string
              format: date-time
              description: When the error occurred (ISO 8601)